Move some paths
