+++ extensions/superboogav2/requirements.txt
@@ -1,7 +1,8 @@
 beautifulsoup4==4.12.2
 chromadb==0.3.18
+lxml
+optuna
+pandas==2.0.3
 posthog==2.4.2
 sentence_transformers==2.2.2
-lxml
 spacy
-optuna
