+++ extensions/superboogav2/utils.py
@@ -2,7 +2,7 @@
 This module contains common functions across multiple other modules.
 """
 
-import extensions.superbooga.parameters as parameters
+import extensions.superboogav2.parameters as parameters
 
 # Create the context using the prefix + data_separator + postfix from parameters.
 def create_context_text(results):
